+++ CHANGELOG.md
@@ -2,3 +2,4 @@ Change Log for Authress Ruby SDK:
 
 ## 2.0
 * Update minimum required ruby version to 3.0
+* Support Omniauth login